huge refactor for font loading | shader path for windows ppl
