+++ client/src/sge/FontLoader.cpp
@@ -1,0 +1,76 @@
+#include "sge/FontLoader.h"
+
+namespace sge {
+
+
+    std::map<char, Character> Characters; 
+
+    void sge::loadFont() {
+        //////// baddy experiment below (builds and run now)//////////
+
+        FT_Library ft;
+        FT_Init_FreeType(&ft);
+        if (FT_Init_FreeType(&ft))
+        {
+            std::cout << "ERROR::FREETYPE: Could not init FreeType Library" << std::endl;
+            // return -1;
+        }
+
+        FT_Face face;
+        std::string fontpath_str = std::string(PROJECT_PATH) + SetupParser::getValue("font-path");
+        const char* fontpath = fontpath_str.c_str();
+        // std::printf("font path %s\n", fontpath);
+        if (FT_New_Face(ft, fontpath, 0, &face))
+        {
+            std::cout << "ERROR::FREETYPE: Failed to load font" << std::endl;  
+            // return -1;
+        }
+        FT_Set_Pixel_Sizes(face, 0, 48);  
+        glPixelStorei(GL_UNPACK_ALIGNMENT, 1); // disable byte-alignment restriction
+    
+        for (unsigned char c = 0; c < 128; c++)
+        {
+            // load character glyph 
+            if (FT_Load_Char(face, c, FT_LOAD_RENDER))
+            {
+                std::cout << "ERROR::FREETYTPE: Failed to load Glyph" << std::endl;
+                continue;
+            }
+            // generate texture
+            unsigned int texture;
+            glGenTextures(1, &texture);
+            glBindTexture(GL_TEXTURE_2D, texture);
+            glTexImage2D(
+                GL_TEXTURE_2D,
+                0,
+                GL_RED,
+                face->glyph->bitmap.width,
+                face->glyph->bitmap.rows,
+                0,
+                GL_RED,
+                GL_UNSIGNED_BYTE,
+                face->glyph->bitmap.buffer
+            );
+            // set texture options
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
+            glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
+            // now store character for later use
+            Character character = {
+                texture, 
+                glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
+                glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
+                static_cast<unsigned int>(face->glyph->advance.x)
+            };
+            Characters.insert(std::pair<char, Character>(c, character));
+        }
+
+        FT_Done_Face(face);
+        FT_Done_FreeType(ft);
+
+        //////// font experiment above ///////
+    }
+
+
+}
